+++ code2inv/prog_generator/fuzz/src/83.c
@@ -66,8 +66,8 @@ void loopcheck(FILE *file_descp, char *buff, int x, int y)
 int main()
 {
   // variable declarations
-  int x;
-  int y;
+  long long int x;
+  long long int y;
 
   char buff[500];
   memset(buff, '\0', sizeof(buff));
@@ -79,7 +79,7 @@ int main()
   for (;;)
   {
     size_t len;
-    const int8_t *buf;
+    const int32_t *buf;
 
     HF_ITER(&buf, &len);
